Moved fsUtil to utils foldes
