+++ src/lib/utils/fsUtil.ts
